Move AI SDK UI adapter up one level
