+++ docs/ai-python/content/docs/reference/ai.agents/meta.json
@@ -1,7 +1,5 @@
 {
   "title": "ai.agents",
   "description": "Reference for advanced agent namespace APIs.",
-  "pages": [
-    "ui"
-  ]
+  "pages": []
 }
