+++ docs/ai-python/content/docs/reference/ai.ui/ai-sdk/approvals.mdx
@@ -1,5 +1,5 @@
 ---
-title: "ai.agents.ui.ai_sdk approvals"
+title: "ai.ui.ai_sdk approvals"
 description: Apply AI SDK UI approval responses.
 type: reference
 summary: Reference for ApprovalResponse, extract_approvals, and apply_approvals.
@@ -15,6 +15,6 @@ registry.
 - `apply_approvals`
 
 ```python
-approvals = ai.agents.ui.ai_sdk.extract_approvals(ui_messages)
-ai.agents.ui.ai_sdk.apply_approvals(approvals)
+approvals = ai.ui.ai_sdk.extract_approvals(ui_messages)
+ai.ui.ai_sdk.apply_approvals(approvals)
 ```
