+++ docs/ai-python/content/docs/reference/ai.ui/ai-sdk/inbound-messages.mdx
@@ -1,5 +1,5 @@
 ---
-title: "ai.agents.ui.ai_sdk.to_messages"
+title: "ai.ui.ai_sdk.to_messages"
 description: Convert UI messages to runtime messages.
 type: reference
 summary: Reference for inbound AI SDK UI message conversion.
@@ -9,7 +9,7 @@ summary: Reference for inbound AI SDK UI message conversion.
 approval responses.
 
 ```python
-messages, approvals = ai.agents.ui.ai_sdk.to_messages(ui_messages)
+messages, approvals = ai.ui.ai_sdk.to_messages(ui_messages)
 ```
 
 Call `apply_approvals` before resuming an agent run with extracted approvals.
