+++ docs/ai-python/content/docs/reference/ai.ui/ai-sdk/index.mdx
@@ -1,11 +1,11 @@
 ---
-title: "ai.agents.ui.ai_sdk"
+title: "ai.ui.ai_sdk"
 description: Reference for AI SDK UI message and SSE adapters.
 type: reference
-summary: Reference for ai.agents.ui.ai_sdk.
+summary: Reference for ai.ui.ai_sdk.
 ---
 
-`ai.agents.ui.ai_sdk` converts between AI SDK UI message streams and the Python
+`ai.ui.ai_sdk` converts between AI SDK UI message streams and the Python
 runtime.
 
 ## APIs
