+++ docs/ai-python/content/docs/reference/ai.ui/ai-sdk/outbound-messages.mdx
@@ -1,14 +1,14 @@
 ---
-title: "ai.agents.ui.ai_sdk.to_ui_messages"
+title: "ai.ui.ai_sdk.to_ui_messages"
 description: Convert runtime messages to AI SDK UI messages.
 type: reference
-summary: Reference for ai.agents.ui.ai_sdk.to_ui_messages.
+summary: Reference for ai.ui.ai_sdk.to_ui_messages.
 ---
 
 `to_ui_messages` converts stored runtime messages back to AI SDK UI messages.
 
 ```python
-ui_messages = ai.agents.ui.ai_sdk.to_ui_messages(messages)
+ui_messages = ai.ui.ai_sdk.to_ui_messages(messages)
 ```
 
 Use it when loading durable conversation history for a UI client.
