+++ src/ai/__init__.py
@@ -1,4 +1,4 @@
-from . import errors, models, providers, telemetry, util
+from . import errors, models, providers, telemetry, ui, util
 from .agents import (
     Agent,
     AgentTool,
@@ -203,6 +203,7 @@ __all__ = [
     "tool_result",
     "tool_result_part",
     "tools",
+    "ui",
     "user_message",
     "util",
     "yield_from",
