+++ src/ai/agents/__init__.py
@@ -1,5 +1,5 @@
 from ..types.messages import MessageBundle
-from . import mcp, ui
+from . import mcp
 from .agent import (
     Agent,
     AgentTool,
@@ -65,6 +65,5 @@ __all__ = [
     "resolve_hook",
     "tool",
     "tool_result",
-    "ui",
     "yield_from",
 ]
